sweeping the floor

+++ Backend/src/main/java/ca/mcgill/ecse321/sportsregistrationw24/service/CourseSessionService.java
@@ -2,11 +2,15 @@ package ca.mcgill.ecse321.sportsregistrationw24.service;
 
 import ca.mcgill.ecse321.sportsregistrationw24.dao.CourseOfferingRepository;
 import ca.mcgill.ecse321.sportsregistrationw24.dao.CourseSessionRepository;
+import ca.mcgill.ecse321.sportsregistrationw24.dao.UserAccountRepository;
 import ca.mcgill.ecse321.sportsregistrationw24.model.CourseOffering;
 import ca.mcgill.ecse321.sportsregistrationw24.model.CourseSession;
 
+import ca.mcgill.ecse321.sportsregistrationw24.model.InstructorAccount;
+import ca.mcgill.ecse321.sportsregistrationw24.model.UserAccount;
 import ca.mcgill.ecse321.sportsregistrationw24.utilities.Utilities;
 import jakarta.transaction.Transactional;
+import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,6 +22,9 @@ import java.time.LocalDate;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
+
+import static ca.mcgill.ecse321.sportsregistrationw24.utilities.Utilities.getUserFromToken;
 
 @Service
 public class CourseSessionService {
@@ -25,10 +32,28 @@ public class CourseSessionService {
     private CourseSessionRepository courseSessionRepository;
     @Autowired
     private CourseOfferingRepository courseOfferingRepository;
+    @Autowired
+    private UserAccountRepository userAccountRepository;
 
     @Transactional
     //Use when you are creating a singular courseSession (Course Offerings where you know for a fact will only have one associated session)
-    public void createCourseSession (Date aDate, Time aStartTime, Time aEndTime, Integer aCourseOfferindId) {
+    public void createCourseSession (String userToken, Date aDate, Time aStartTime, Time aEndTime, Integer aCourseOfferingId) {
+        UserAccount user = getUserFromToken(userAccountRepository, userToken);
+
+        if (!user.getUserType().equals("INSTRUCTOR")) {
+            throw new IllegalArgumentException("Only instructors can create course sessions");
+        }
+
+        CourseOffering courseOffering = courseOfferingRepository.findById(aCourseOfferingId).orElse(null);
+
+        if (courseOffering == null) {
+            throw new IllegalArgumentException("Course Offering not found");
+        }
+
+        if (!courseOffering.getInstructorAccount().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("Instructor does not have permission to create course session for this course offering");
+        }
+
         // NULL POINTER CHECKS
         if (aDate == null) {
             throw new IllegalArgumentException("Date field cannot be null");
@@ -36,8 +61,6 @@ public class CourseSessionService {
             throw new IllegalArgumentException("Start time field cannot be null");
         } else if (aEndTime == null) {
             throw new IllegalArgumentException("End time field cannot be null");
-        } else if (aCourseOfferindId == null) {
-            throw new IllegalArgumentException("Course offering ID field cannot be null");
         }
         // START AND END TIME CHECKS
         if (aStartTime.compareTo(aEndTime) == 0) {
@@ -45,24 +68,11 @@ public class CourseSessionService {
         } else if (aStartTime.compareTo(aEndTime) > 0) {
             throw new IllegalArgumentException("Session start time cannot be after session end time");
         }
-        // SESSION DURATION CHECK (CANNOT BE OVER 1 HOUR)
-        /**
-         long sessionDuration = Math.abs(aStartTime.getTime() - aEndTime.getTime());
-         long oneHour = 60 * 60 * 1000;
-         if (sessionDuration > oneHour) {
-         throw new IllegalArgumentException("Course sessions cannot last for over an hour");
-         }
-         **/
 
-        CourseOffering foundCourseOffering = courseOfferingRepository.findById(aCourseOfferindId).orElse(null);
-
-        if (foundCourseOffering == null) {
-            throw new IllegalArgumentException("Course Offering not found");
-        }
         // DATE CHECKS
-        if (aDate.toLocalDate().isBefore(foundCourseOffering.getStartDate().toLocalDate())) {
+        if (aDate.toLocalDate().isBefore(courseOffering.getStartDate().toLocalDate())) {
             throw new IllegalArgumentException("Date of course session cannot be before start date of course offering");
-        } else if (aDate.toLocalDate().isAfter(foundCourseOffering.getEndDate().toLocalDate())) {
+        } else if (aDate.toLocalDate().isAfter(courseOffering.getEndDate().toLocalDate())) {
             throw new IllegalArgumentException("Date of course session cannot be after end date of course offering");
         }
 
@@ -71,29 +81,39 @@ public class CourseSessionService {
         newCourseSession.setDate(aDate);
         newCourseSession.setStartTime(aStartTime);
         newCourseSession.setEndTime(aEndTime);
-        newCourseSession.setCourseOffering(foundCourseOffering);
+        newCourseSession.setCourseOffering(courseOffering);
 
         courseSessionRepository.save(newCourseSession);
     }
 
     @Transactional
     //Use when you are creating courseSessions from a courseOffering with recurring sessions
-    public void createCourseSessions (HashMap<DayOfWeek, ArrayList<Time>> dayTimeMapping, Integer aCourseOfferingID) {
-        if (dayTimeMapping == null) {
-            throw new IllegalArgumentException("Day time mapping field cannot be null");
-        } else if (aCourseOfferingID == null) {
-            throw new IllegalArgumentException("Course offering ID field cannot be null");
+    public void createCourseSessions (String userToken, HashMap<DayOfWeek, ArrayList<Time>> dayTimeMapping, Integer aCourseOfferingID) {
+        UserAccount user = getUserFromToken(userAccountRepository, userToken);
+
+        if (!user.getUserType().equals("INSTRUCTOR")) {
+            throw new IllegalArgumentException("Only instructors can create course sessions");
         }
 
-        CourseOffering foundCourseOffering = courseOfferingRepository.findById(aCourseOfferingID).orElse(null);
+        CourseOffering courseOffering = courseOfferingRepository.findById(aCourseOfferingID).orElse(null);
 
-        if (foundCourseOffering == null) {
+        if (courseOffering == null) {
             throw new IllegalArgumentException("Course Offering not found");
         }
 
-        LocalDate startDate = foundCourseOffering.getStartDate().toLocalDate();
-        LocalDate endDate = foundCourseOffering.getEndDate().toLocalDate();
-        ArrayList<DayOfWeek> daysOffered = foundCourseOffering.getDaysOffered();
+        if (!courseOffering.getInstructorAccount().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("Instructor does not have permission to create course session for this course offering");
+        }
+
+
+        if (dayTimeMapping == null) {
+            throw new IllegalArgumentException("Day time mapping field cannot be null");
+        }
+
+
+        LocalDate startDate = courseOffering.getStartDate().toLocalDate();
+        LocalDate endDate = courseOffering.getEndDate().toLocalDate();
+        ArrayList<DayOfWeek> daysOffered = courseOffering.getDaysOffered();
 
         LocalDate generatedDate = startDate;
         ArrayList<Date> sessionDates = new ArrayList<Date>();
@@ -111,7 +131,7 @@ public class CourseSessionService {
             newCourseSession.setDate(sessionDate);
             newCourseSession.setStartTime(dayTimeMapping.get(sessionDate.toLocalDate().getDayOfWeek()).get(0));
             newCourseSession.setEndTime(dayTimeMapping.get(sessionDate.toLocalDate().getDayOfWeek()).get(1));
-            newCourseSession.setCourseOffering(foundCourseOffering);
+            newCourseSession.setCourseOffering(courseOffering);
 
             courseSessionRepository.save(newCourseSession);
         }
@@ -137,10 +157,24 @@ public class CourseSessionService {
     }
 
     @Transactional
-    public ArrayList<CourseSession> getAllCourseSessions () {
-        return Utilities.iterableToArrayList(courseSessionRepository.findAll());
+    public List<CourseSession> getAllCourseSessions (String userToken, Date lowDate, Date highDate, Time lowTime, Time highTime, Integer instructorId) {
+        UserAccount user = getUserFromToken(userAccountRepository, userToken);
+
+        if (!user.getUserType().equals("OWNER")) {
+            throw new IllegalArgumentException("Only owners can view all course sessions");
+        }
+        InstructorAccount instructor = (InstructorAccount) userAccountRepository.findById(instructorId).orElse(null);
+
+        List<CourseSession> foundCourseSessions = courseSessionRepository.findByFilters(lowDate, highDate, lowTime, highTime, instructor).orElse(null);
+
+        if(foundCourseSessions == null){
+            throw new IllegalArgumentException("No course sessions found with the provided information");
+        }
+        return foundCourseSessions;
     }
 
+    // when would we ever use this?
+    @Deprecated
     @Transactional
     public CourseSession getCourseSession(Integer courseSessionId) {
         if (courseSessionId == null) {
@@ -161,7 +195,7 @@ public class CourseSessionService {
     }
 
     @Transactional
-    public ArrayList<CourseSession> getCourseSessionsByCourseOfferingID (Integer courseOfferingID) {
+    public List<CourseSession> getCourseSessionsByCourseOfferingID (Integer courseOfferingID) {
         CourseOffering foundCourseOffering = courseOfferingRepository.findById(courseOfferingID).orElse(null);
 
         if (foundCourseOffering == null) {
@@ -178,23 +212,54 @@ public class CourseSessionService {
     }
 
     @Transactional
-    public void deleteCourseSessionByID (Integer aID) {
+    public void deleteCourseSessionByID (String userToken, Integer aID) {
+        UserAccount user = getUserFromToken(userAccountRepository, userToken);
+
+        if (user.getUserType().equals("CUSTOMER")) { //when would this ever happen??!
+            throw new IllegalArgumentException("Customers cannot delete course sessions");
+        }
+
+        CourseSession foundCourseSession = courseSessionRepository.findById(aID).orElse(null);
+
+        if (foundCourseSession == null) {
+            throw new IllegalArgumentException("Course Session not found");
+        }
+
+        if (user.getUserType().equals("INSTRUCTOR")) {
+            if (!foundCourseSession.getCourseOffering().getInstructorAccount().getId().equals(user.getId())) {
+                throw new IllegalArgumentException("Instructor does not have permission to delete this course session");
+            }
+
+        }
+
         courseSessionRepository.deleteById(aID);
     }
 
 
     @Transactional
-    public void deleteCourseSessionsByCourseOfferingID (Integer courseOfferingID) {
+    public void deleteCourseSessionsByCourseOfferingID (String userToken, Integer courseOfferingID) {
+        UserAccount user = getUserFromToken(userAccountRepository, userToken);
+
+        if (user.getUserType().equals("CUSTOMER")) {
+            throw new IllegalArgumentException("Customers cannot delete course sessions");
+        }
+
         CourseOffering foundCourseOffering = courseOfferingRepository.findById(courseOfferingID).orElse(null);
 
         if (foundCourseOffering == null) {
             throw new IllegalArgumentException("Course Offering not found");
         }
 
+        if (user.getUserType().equals("INSTRUCTOR")) {
+            if (!foundCourseOffering.getInstructorAccount().getId().equals(user.getId())) {
+                throw new IllegalArgumentException("Instructor does not have permission to delete course sessions for this course offering");
+            }
+        }
+
         Iterable<CourseSession> foundCourseSessions = courseSessionRepository.findByCourseOffering(foundCourseOffering).orElse(null);
 
         if (foundCourseSessions == null) {
-            return;
+            throw new IllegalArgumentException("No course sessions found for this course offering");
         }
 
         for (CourseSession foundCourseSession : foundCourseSessions) {
